backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
pymongo==4.6.1
vaderSentiment==3.3.2
tweepy==4.14.0
python-dotenv==1.0.0
eventlet==0.33.3
